[PATCH 02/67] Force Mesa to use the PVR driver for platform devices
For platform devices, Mesa uses the DRM driver name to determine which DRI
driver to load. This doesn't work in the multi-DRM driver model where there
are separate DRM drivers for the display and GPU. This is because we normally
want Mesa to deal with the display driver and the name of this won't match
that of the DRI driver. For the time being, take the easy approach and just
force it to use the PVR DRI driver.