egl/null: add support for DRI_PRIME GPU selection
Add support for selecting the GPU to be used for rendering using
the DRI_PRIME environment variable. If a different GPU is selected,
a duplicate of the file descriptor for the original GPU/display is
preserved, which can be obtained by calling the getDisplayFD
function in the image loader extension.

This change includes code, in function dri2_null_try_device, to
skip display devices that are not supported by the PVR driver.
This is to prevent failure on PC based test systems, that often
include display hardware that is incompatible with the driver.
This code would not be needed for systems that don't use the PVR
driver, or on production systems that use the PVR driver.